@@ -1,8 +1,7 @@
-﻿namespace Main.Data.Models
+﻿namespace Data.Models;
+
+public class GroupModel : BaseModel<Guid>
 {
-    public class GroupModel : BaseModel<Guid>
-    {
-        public required string Name { get; set; } = string.Empty;
-        public string? Descriere { get; set; } = null;
-    }
+    public required string Name { get; set; }
+    public string? Description { get; set; } = null;
 }
