@@ -1,13 +1,12 @@
-namespace Main.Data.Models
+namespace Data.Models;
+
+public class LinkModel : BaseModel<Guid>
 {
-    public class LinkModel : BaseModel<Guid>
-    {
-        public string SourceType { get; set; } = string.Empty; // e.g., "User"
-        public Guid SourceId { get; set; }
+    public string SourceType { get; set; } = string.Empty; // e.g., "User"
+    public Guid SourceId { get; set; }
 
-        public string TargetType { get; set; } = string.Empty; // e.g., "UserRole"
-        public Guid TargetId { get; set; }
+    public string TargetType { get; set; } = string.Empty; // e.g., "UserRole"
+    public Guid TargetId { get; set; }
 
-        public string? Metadata { get; set; } = null;
-    }
+    public string? Metadata { get; set; } = null;
 }
