@@ -1,9 +1,13 @@
 /* tslint:disable */
 /* eslint-disable */
-export * from './AdmApi';
 export * from './AuthApi';
-export * from './CommunityApi';
+export * from './ChatApi';
 export * from './DriveApi';
+export * from './FriendsApi';
+export * from './GroupsApi';
+export * from './LinksApi';
+export * from './NotificationsApi';
+export * from './ParametersApi';
 export * from './PostsApi';
 export * from './QuestionAnswersApi';
 export * from './UsersApi';
