@@ -1,6 +1,5 @@
-// Status: Unknowm(Perfect)
-
 import { app } from "@/main";
+import type { App } from "vue";
 
 export const apiRequest = (
   actionPath: string,
@@ -12,8 +11,8 @@ export const apiRequest = (
       bodyData instanceof FormData
         ? ""
         : method == "GET"
-        ? "text/plain"
-        : "application/json",
+          ? "text/plain"
+          : "application/json",
   }
 ) =>
   fetch(`/api/${actionPath}${queryParams(bodyData)}`, {
@@ -31,8 +30,8 @@ export const apiRequest = (
     const content = contentType?.includes("application/json")
       ? await response.json()
       : contentType?.includes("text/plain")
-      ? await response.text()
-      : await response.blob();
+        ? await response.text()
+        : await response.blob();
 
     app.config.globalProperties.$toast?.add({
       summary: "Server",
@@ -40,8 +39,9 @@ export const apiRequest = (
       severity: response.ok
         ? "success"
         : response.status >= 400 && response.status < 600
-        ? "error"
-        : "info",
+          ? "error"
+          : "info",
+      life: 3000,
     });
     console.log(content);
     return content;
@@ -51,3 +51,51 @@ export const queryParams = (bodyData: any) =>
   Object.values(bodyData).every((value) => value != undefined && value != "")
     ? "?" + new URLSearchParams(bodyData)
     : "";
+
+export function setupGlobalFetchInterceptor(appInstance: App) {
+  const originalFetch = window.fetch;
+
+  window.fetch = async function (
+    input: RequestInfo | URL,
+    init: RequestInit = {}
+  ): Promise<Response> {
+    const forcedHeaders: HeadersInit = {
+      Authorization: `Bearer ${getCookie("access_token")}`,
+      "X-Requested-With": "XMLHttpRequest",
+    };
+
+    const headers: HeadersInit = {
+      ...(init.headers || {}),
+      ...forcedHeaders,
+    };
+
+    const response = await originalFetch(input, { ...init, headers });
+
+    const clone = response.clone();
+    let message = "";
+
+    try {
+      const text = await clone.text();
+      message = text?.length > 200 ? text.slice(0, 200) + "..." : text;
+    } catch {
+      message = "[Unable to read response body]";
+    }
+
+    const severity = response.ok
+      ? "success"
+      : response.status >= 400 && response.status < 600
+        ? "error"
+        : "info";
+
+    if (message && appInstance.config.globalProperties.$toast) {
+      appInstance.config.globalProperties.$toast.add({
+        summary: capitalizeWords(severity),
+        detail: message || "No response body",
+        severity,
+        life: 3000,
+      });
+    }
+
+    return response;
+  };
+}
