@@ -5,76 +5,71 @@
  * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
  *
  * The version of the OpenAPI document: 1.0.0
- *
+ * 
  *
  * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
  * https://openapi-generator.tech
  * Do not edit the class manually.
  */
 
-import { mapValues } from "../runtime";
+import { mapValues } from '../runtime';
 /**
- *
+ * 
  * @export
  * @interface AuthenticationDTO
  */
 export interface AuthenticationDTO {
-  /**
-   *
-   * @type {string}
-   * @memberof AuthenticationDTO
-   */
-  username: string;
-  /**
-   *
-   * @type {string}
-   * @memberof AuthenticationDTO
-   */
-  password: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof AuthenticationDTO
+     */
+    username: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof AuthenticationDTO
+     */
+    password: string;
 }
 
 /**
  * Check if a given object implements the AuthenticationDTO interface.
  */
-export function instanceOfAuthenticationDTO(
-  value: object
-): value is AuthenticationDTO {
-  if (!("username" in value) || value["username"] == undefined) return false;
-  if (!("password" in value) || value["password"] == undefined) return false;
-  return true;
+export function instanceOfAuthenticationDTO(value: object): value is AuthenticationDTO {
+    if (!('username' in value) || value['username'] === undefined) return false;
+    if (!('password' in value) || value['password'] === undefined) return false;
+    return true;
 }
 
 export function AuthenticationDTOFromJSON(json: any): AuthenticationDTO {
-  return AuthenticationDTOFromJSONTyped(json, false);
+    return AuthenticationDTOFromJSONTyped(json, false);
 }
 
-export function AuthenticationDTOFromJSONTyped(
-  json: any,
-  ignoreDiscriminator: boolean
-): AuthenticationDTO {
-  if (json == null) {
-    return json;
-  }
-  return {
-    username: json["username"],
-    password: json["password"],
-  };
+export function AuthenticationDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthenticationDTO {
+    if (json == null) {
+        return json;
+    }
+    return {
+        
+        'username': json['username'],
+        'password': json['password'],
+    };
 }
 
 export function AuthenticationDTOToJSON(json: any): AuthenticationDTO {
-  return AuthenticationDTOToJSONTyped(json, false);
+    return AuthenticationDTOToJSONTyped(json, false);
 }
 
-export function AuthenticationDTOToJSONTyped(
-  value?: AuthenticationDTO | null,
-  ignoreDiscriminator: boolean = false
-): any {
-  if (value == null) {
-    return value;
-  }
+export function AuthenticationDTOToJSONTyped(value?: AuthenticationDTO | null, ignoreDiscriminator: boolean = false): any {
+    if (value == null) {
+        return value;
+    }
 
-  return {
-    username: value["username"],
-    password: value["password"],
-  };
+    return {
+        
+        'username': value['username'],
+        'password': value['password'],
+    };
 }
+
