@@ -5,106 +5,106 @@
  * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
  *
  * The version of the OpenAPI document: 1.0.0
- *
+ * 
  *
  * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
  * https://openapi-generator.tech
  * Do not edit the class manually.
  */
 
-import { mapValues } from "../runtime";
+import { mapValues } from '../runtime';
+import type { ApiDriveUploadPostRequestLength } from './ApiDriveUploadPostRequestLength';
+import {
+    ApiDriveUploadPostRequestLengthFromJSON,
+    ApiDriveUploadPostRequestLengthFromJSONTyped,
+    ApiDriveUploadPostRequestLengthToJSON,
+    ApiDriveUploadPostRequestLengthToJSONTyped,
+} from './ApiDriveUploadPostRequestLength';
+
 /**
- *
+ * 
  * @export
  * @interface DriveInfoDTO
  */
 export interface DriveInfoDTO {
-  /**
-   *
-   * @type {string}
-   * @memberof DriveInfoDTO
-   */
-  name: string;
-  /**
-   *
-   * @type {string}
-   * @memberof DriveInfoDTO
-   */
-  volumeLabel: string;
-  /**
-   *
-   * @type {string}
-   * @memberof DriveInfoDTO
-   */
-  driveType: string;
-  /**
-   *
-   * @type {number}
-   * @memberof DriveInfoDTO
-   */
-  availableFreeSpace: number;
-  /**
-   *
-   * @type {number}
-   * @memberof DriveInfoDTO
-   */
-  totalSize: number;
+    /**
+     * 
+     * @type {string}
+     * @memberof DriveInfoDTO
+     */
+    name: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof DriveInfoDTO
+     */
+    volumeLabel: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof DriveInfoDTO
+     */
+    driveType: string;
+    /**
+     * 
+     * @type {ApiDriveUploadPostRequestLength}
+     * @memberof DriveInfoDTO
+     */
+    availableFreeSpace: ApiDriveUploadPostRequestLength;
+    /**
+     * 
+     * @type {ApiDriveUploadPostRequestLength}
+     * @memberof DriveInfoDTO
+     */
+    totalSize: ApiDriveUploadPostRequestLength;
 }
 
 /**
  * Check if a given object implements the DriveInfoDTO interface.
  */
 export function instanceOfDriveInfoDTO(value: object): value is DriveInfoDTO {
-  if (!("name" in value) || value["name"] == undefined) return false;
-  if (!("volumeLabel" in value) || value["volumeLabel"] == undefined)
-    return false;
-  if (!("driveType" in value) || value["driveType"] == undefined) return false;
-  if (
-    !("availableFreeSpace" in value) ||
-    value["availableFreeSpace"] == undefined
-  )
-    return false;
-  if (!("totalSize" in value) || value["totalSize"] == undefined) return false;
-  return true;
+    if (!('name' in value) || value['name'] === undefined) return false;
+    if (!('volumeLabel' in value) || value['volumeLabel'] === undefined) return false;
+    if (!('driveType' in value) || value['driveType'] === undefined) return false;
+    if (!('availableFreeSpace' in value) || value['availableFreeSpace'] === undefined) return false;
+    if (!('totalSize' in value) || value['totalSize'] === undefined) return false;
+    return true;
 }
 
 export function DriveInfoDTOFromJSON(json: any): DriveInfoDTO {
-  return DriveInfoDTOFromJSONTyped(json, false);
+    return DriveInfoDTOFromJSONTyped(json, false);
 }
 
-export function DriveInfoDTOFromJSONTyped(
-  json: any,
-  ignoreDiscriminator: boolean
-): DriveInfoDTO {
-  if (json == null) {
-    return json;
-  }
-  return {
-    name: json["name"],
-    volumeLabel: json["volumeLabel"],
-    driveType: json["driveType"],
-    availableFreeSpace: json["availableFreeSpace"],
-    totalSize: json["totalSize"],
-  };
+export function DriveInfoDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): DriveInfoDTO {
+    if (json == null) {
+        return json;
+    }
+    return {
+        
+        'name': json['name'],
+        'volumeLabel': json['volumeLabel'],
+        'driveType': json['driveType'],
+        'availableFreeSpace': ApiDriveUploadPostRequestLengthFromJSON(json['availableFreeSpace']),
+        'totalSize': ApiDriveUploadPostRequestLengthFromJSON(json['totalSize']),
+    };
 }
 
 export function DriveInfoDTOToJSON(json: any): DriveInfoDTO {
-  return DriveInfoDTOToJSONTyped(json, false);
+    return DriveInfoDTOToJSONTyped(json, false);
 }
 
-export function DriveInfoDTOToJSONTyped(
-  value?: DriveInfoDTO | null,
-  ignoreDiscriminator: boolean = false
-): any {
-  if (value == null) {
-    return value;
-  }
+export function DriveInfoDTOToJSONTyped(value?: DriveInfoDTO | null, ignoreDiscriminator: boolean = false): any {
+    if (value == null) {
+        return value;
+    }
 
-  return {
-    name: value["name"],
-    volumeLabel: value["volumeLabel"],
-    driveType: value["driveType"],
-    availableFreeSpace: value["availableFreeSpace"],
-    totalSize: value["totalSize"],
-  };
+    return {
+        
+        'name': value['name'],
+        'volumeLabel': value['volumeLabel'],
+        'driveType': value['driveType'],
+        'availableFreeSpace': ApiDriveUploadPostRequestLengthToJSON(value['availableFreeSpace']),
+        'totalSize': ApiDriveUploadPostRequestLengthToJSON(value['totalSize']),
+    };
 }
+
