@@ -5,194 +5,173 @@
  * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
  *
  * The version of the OpenAPI document: 1.0.0
- *
+ * 
  *
  * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
  * https://openapi-generator.tech
  * Do not edit the class manually.
  */
 
-import { mapValues } from "../runtime";
+import { mapValues } from '../runtime';
 /**
- *
+ * 
  * @export
  * @interface UserModel
  */
 export interface UserModel {
-  /**
-   *
-   * @type {string}
-   * @memberof UserModel
-   */
-  username?: string;
-  /**
-   *
-   * @type {string}
-   * @memberof UserModel
-   */
-  password?: string;
-  /**
-   *
-   * @type {string}
-   * @memberof UserModel
-   */
-  email?: string;
-  /**
-   *
-   * @type {string}
-   * @memberof UserModel
-   */
-  signiture?: string | null;
-  /**
-   *
-   * @type {Date}
-   * @memberof UserModel
-   */
-  birthDay?: Date | null;
-  /**
-   *
-   * @type {Array<string>}
-   * @memberof UserModel
-   */
-  friends?: Array<string>;
-  /**
-   *
-   * @type {Array<string>}
-   * @memberof UserModel
-   */
-  saves?: Array<string>;
-  /**
-   *
-   * @type {string}
-   * @memberof UserModel
-   */
-  name?: string;
-  /**
-   *
-   * @type {Date}
-   * @memberof UserModel
-   */
-  emailVerifiedDate?: Date | null;
-  /**
-   *
-   * @type {string}
-   * @memberof UserModel
-   */
-  avatar?: string | null;
-  /**
-   *
-   * @type {string}
-   * @memberof UserModel
-   */
-  id?: string;
-  /**
-   *
-   * @type {Date}
-   * @memberof UserModel
-   */
-  executionDate?: Date | null;
-  /**
-   *
-   * @type {string}
-   * @memberof UserModel
-   */
-  authorId?: string | null;
-  /**
-   *
-   * @type {UserModel}
-   * @memberof UserModel
-   */
-  authorData?: UserModel | null;
-  /**
-   *
-   * @type {Array<string>}
-   * @memberof UserModel
-   */
-  tags?: Array<string> | null;
+    /**
+     * 
+     * @type {string}
+     * @memberof UserModel
+     */
+    username?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof UserModel
+     */
+    password?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof UserModel
+     */
+    email?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof UserModel
+     */
+    signiture?: string | null;
+    /**
+     * 
+     * @type {Date}
+     * @memberof UserModel
+     */
+    birthDay?: Date | null;
+    /**
+     * 
+     * @type {Array<string>}
+     * @memberof UserModel
+     */
+    friends?: Array<string>;
+    /**
+     * 
+     * @type {Array<string>}
+     * @memberof UserModel
+     */
+    saves?: Array<string>;
+    /**
+     * 
+     * @type {string}
+     * @memberof UserModel
+     */
+    name?: string;
+    /**
+     * 
+     * @type {Date}
+     * @memberof UserModel
+     */
+    emailVerifiedDate?: Date | null;
+    /**
+     * 
+     * @type {string}
+     * @memberof UserModel
+     */
+    avatar?: string | null;
+    /**
+     * 
+     * @type {string}
+     * @memberof UserModel
+     */
+    id?: string;
+    /**
+     * 
+     * @type {Date}
+     * @memberof UserModel
+     */
+    executionDate?: Date | null;
+    /**
+     * 
+     * @type {string}
+     * @memberof UserModel
+     */
+    authorId?: string | null;
+    /**
+     * 
+     * @type {UserModel}
+     * @memberof UserModel
+     */
+    authorData?: UserModel | null;
+    /**
+     * 
+     * @type {Array<string>}
+     * @memberof UserModel
+     */
+    tags?: Array<string> | null;
 }
 
 /**
  * Check if a given object implements the UserModel interface.
  */
 export function instanceOfUserModel(value: object): value is UserModel {
-  return true;
+    return true;
 }
 
 export function UserModelFromJSON(json: any): UserModel {
-  return UserModelFromJSONTyped(json, false);
+    return UserModelFromJSONTyped(json, false);
 }
 
-export function UserModelFromJSONTyped(
-  json: any,
-  ignoreDiscriminator: boolean
-): UserModel {
-  if (json == null) {
-    return json;
-  }
-  return {
-    username: json["username"] == null ? undefined : json["username"],
-    password: json["password"] == null ? undefined : json["password"],
-    email: json["email"] == null ? undefined : json["email"],
-    signiture: json["signiture"] == null ? undefined : json["signiture"],
-    birthDay: json["birthDay"] == null ? undefined : new Date(json["birthDay"]),
-    friends: json["friends"] == null ? undefined : json["friends"],
-    saves: json["saves"] == null ? undefined : json["saves"],
-    name: json["name"] == null ? undefined : json["name"],
-    emailVerifiedDate:
-      json["emailVerifiedDate"] == null
-        ? undefined
-        : new Date(json["emailVerifiedDate"]),
-    avatar: json["avatar"] == null ? undefined : json["avatar"],
-    id: json["id"] == null ? undefined : json["id"],
-    executionDate:
-      json["executionDate"] == null
-        ? undefined
-        : new Date(json["executionDate"]),
-    authorId: json["authorId"] == null ? undefined : json["authorId"],
-    authorData:
-      json["authorData"] == null
-        ? undefined
-        : UserModelFromJSON(json["authorData"]),
-    tags: json["tags"] == null ? undefined : json["tags"],
-  };
+export function UserModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserModel {
+    if (json == null) {
+        return json;
+    }
+    return {
+        
+        'username': json['username'] == null ? undefined : json['username'],
+        'password': json['password'] == null ? undefined : json['password'],
+        'email': json['email'] == null ? undefined : json['email'],
+        'signiture': json['signiture'] == null ? undefined : json['signiture'],
+        'birthDay': json['birthDay'] == null ? undefined : (new Date(json['birthDay'])),
+        'friends': json['friends'] == null ? undefined : json['friends'],
+        'saves': json['saves'] == null ? undefined : json['saves'],
+        'name': json['name'] == null ? undefined : json['name'],
+        'emailVerifiedDate': json['emailVerifiedDate'] == null ? undefined : (new Date(json['emailVerifiedDate'])),
+        'avatar': json['avatar'] == null ? undefined : json['avatar'],
+        'id': json['id'] == null ? undefined : json['id'],
+        'executionDate': json['executionDate'] == null ? undefined : (new Date(json['executionDate'])),
+        'authorId': json['authorId'] == null ? undefined : json['authorId'],
+        'authorData': json['authorData'] == null ? undefined : UserModelFromJSON(json['authorData']),
+        'tags': json['tags'] == null ? undefined : json['tags'],
+    };
 }
 
 export function UserModelToJSON(json: any): UserModel {
-  return UserModelToJSONTyped(json, false);
+    return UserModelToJSONTyped(json, false);
 }
 
-export function UserModelToJSONTyped(
-  value?: UserModel | null,
-  ignoreDiscriminator: boolean = false
-): any {
-  if (value == null) {
-    return value;
-  }
+export function UserModelToJSONTyped(value?: UserModel | null, ignoreDiscriminator: boolean = false): any {
+    if (value == null) {
+        return value;
+    }
 
-  return {
-    username: value["username"],
-    password: value["password"],
-    email: value["email"],
-    signiture: value["signiture"],
-    birthDay:
-      value["birthDay"] == null
-        ? null
-        : (value["birthDay"] as any)?.toISOString(),
-    friends: value["friends"],
-    saves: value["saves"],
-    name: value["name"],
-    emailVerifiedDate:
-      value["emailVerifiedDate"] == null
-        ? null
-        : (value["emailVerifiedDate"] as any)?.toISOString(),
-    avatar: value["avatar"],
-    id: value["id"],
-    executionDate:
-      value["executionDate"] == null
-        ? null
-        : (value["executionDate"] as any)?.toISOString(),
-    authorId: value["authorId"],
-    authorData: UserModelToJSON(value["authorData"]),
-    tags: value["tags"],
-  };
+    return {
+        
+        'username': value['username'],
+        'password': value['password'],
+        'email': value['email'],
+        'signiture': value['signiture'],
+        'birthDay': value['birthDay'] == null ? value['birthDay'] : value['birthDay'].toISOString(),
+        'friends': value['friends'],
+        'saves': value['saves'],
+        'name': value['name'],
+        'emailVerifiedDate': value['emailVerifiedDate'] == null ? value['emailVerifiedDate'] : value['emailVerifiedDate'].toISOString(),
+        'avatar': value['avatar'],
+        'id': value['id'],
+        'executionDate': value['executionDate'] == null ? value['executionDate'] : value['executionDate'].toISOString(),
+        'authorId': value['authorId'],
+        'authorData': UserModelToJSON(value['authorData']),
+        'tags': value['tags'],
+    };
 }
+
