@@ -1,7 +1,7 @@
 ﻿// <auto-generated />
 using System;
 using System.Collections.Generic;
-using Main.Data.Context;
+using Data.Context;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Migrations;
@@ -10,10 +10,10 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 
 #nullable disable
 
-namespace Main.Data.Migrations
+namespace Data.Migrations
 {
     [DbContext(typeof(BaseContext))]
-    [Migration("20251104165738_1")]
+    [Migration("20251107200315_1")]
     partial class _1
     {
         /// <inheritdoc />
@@ -27,7 +27,7 @@ namespace Main.Data.Migrations
 
             NpgsqlModelBuilderExtensions.UseIdentityByDefaultColumns(modelBuilder);
 
-            modelBuilder.Entity("Main.Data.Models.GroupModel", b =>
+            modelBuilder.Entity("Data.Models.GroupModel", b =>
                 {
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd()
@@ -39,7 +39,7 @@ namespace Main.Data.Migrations
                     b.Property<Guid?>("AuthorId")
                         .HasColumnType("uuid");
 
-                    b.Property<string>("Descriere")
+                    b.Property<string>("Description")
                         .HasColumnType("text");
 
                     b.Property<DateTimeOffset?>("ExecutionDate")
@@ -59,7 +59,7 @@ namespace Main.Data.Migrations
                     b.ToTable("Groups", "base");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.LinkModel", b =>
+            modelBuilder.Entity("Data.Models.LinkModel", b =>
                 {
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd()
@@ -105,7 +105,7 @@ namespace Main.Data.Migrations
                     b.ToTable("Links", "base");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.NotificationModel", b =>
+            modelBuilder.Entity("Data.Models.MessageModel", b =>
                 {
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd()
@@ -120,6 +120,11 @@ namespace Main.Data.Migrations
                     b.Property<Guid?>("AuthorId")
                         .HasColumnType("uuid");
 
+                    b.Property<string>("Discriminator")
+                        .IsRequired()
+                        .HasMaxLength(13)
+                        .HasColumnType("character varying(13)");
+
                     b.Property<DateTimeOffset?>("ExecutionDate")
                         .HasColumnType("timestamptz");
 
@@ -127,27 +132,31 @@ namespace Main.Data.Migrations
                         .IsRequired()
                         .HasColumnType("text");
 
-                    b.Property<DateTimeOffset?>("ReadDate")
+                    b.Property<DateTimeOffset?>("SeenDate")
                         .HasColumnType("timestamptz");
 
                     b.PrimitiveCollection<List<string>>("Tags")
                         .HasColumnType("text[]");
+
+                    b.Property<Guid>("To")
+                        .HasColumnType("uuid");
 
                     b.HasKey("Id");
 
                     b.HasIndex("AuthorDataId");
 
                     b.ToTable("Notifications", "base");
+
+                    b.HasDiscriminator().HasValue("MessageModel");
+
+                    b.UseTphMappingStrategy();
                 });
 
-            modelBuilder.Entity("Main.Data.Models.PostModel", b =>
+            modelBuilder.Entity("Data.Models.ParameterModel", b =>
                 {
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd()
                         .HasColumnType("uuid");
-
-                    b.PrimitiveCollection<List<Guid>>("Attachments")
-                        .HasColumnType("uuid[]");
 
                     b.Property<Guid?>("AuthorDataId")
                         .HasColumnType("uuid");
@@ -155,30 +164,27 @@ namespace Main.Data.Migrations
                     b.Property<Guid?>("AuthorId")
                         .HasColumnType("uuid");
 
+                    b.Property<string>("Description")
+                        .HasColumnType("text");
+
                     b.Property<DateTimeOffset?>("ExecutionDate")
                         .HasColumnType("timestamptz");
 
-                    b.Property<string>("Message")
+                    b.Property<string>("Name")
                         .IsRequired()
                         .HasColumnType("text");
 
-                    b.Property<DateTimeOffset?>("SentDate")
-                        .HasColumnType("timestamp with time zone");
-
                     b.PrimitiveCollection<List<string>>("Tags")
                         .HasColumnType("text[]");
-
-                    b.Property<Guid?>("To")
-                        .HasColumnType("uuid");
 
                     b.HasKey("Id");
 
                     b.HasIndex("AuthorDataId");
 
-                    b.ToTable("Posts", "base");
+                    b.ToTable("Parameters", "base");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.QuestionModel", b =>
+            modelBuilder.Entity("Data.Models.QuestionModel", b =>
                 {
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd()
@@ -216,7 +222,7 @@ namespace Main.Data.Migrations
                     b.ToTable("Questions", "base");
                 });
 
-            modelBuilder.Entity("Main.Modules.Audit.AuditModel", b =>
+            modelBuilder.Entity("Modules.Audit.AuditModel", b =>
                 {
                     b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
@@ -258,7 +264,7 @@ namespace Main.Data.Migrations
                     b.ToTable("Audit", "base");
                 });
 
-            modelBuilder.Entity("Main.Modules.Auth.UserModel", b =>
+            modelBuilder.Entity("Modules.Auth.UserModel", b =>
                 {
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd()
@@ -319,63 +325,70 @@ namespace Main.Data.Migrations
                     b.ToTable("Users", "base");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.GroupModel", b =>
+            modelBuilder.Entity("Data.Models.PostModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasBaseType("Data.Models.MessageModel");
+
+                    b.HasDiscriminator().HasValue("PostModel");
+                });
+
+            modelBuilder.Entity("Data.Models.GroupModel", b =>
+                {
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
                     b.Navigation("AuthorData");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.LinkModel", b =>
+            modelBuilder.Entity("Data.Models.LinkModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
                     b.Navigation("AuthorData");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.NotificationModel", b =>
+            modelBuilder.Entity("Data.Models.MessageModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
                     b.Navigation("AuthorData");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.PostModel", b =>
+            modelBuilder.Entity("Data.Models.ParameterModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
                     b.Navigation("AuthorData");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.QuestionModel", b =>
+            modelBuilder.Entity("Data.Models.QuestionModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
                     b.Navigation("AuthorData");
                 });
 
-            modelBuilder.Entity("Main.Modules.Audit.AuditModel", b =>
+            modelBuilder.Entity("Modules.Audit.AuditModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
                     b.Navigation("AuthorData");
                 });
 
-            modelBuilder.Entity("Main.Modules.Auth.UserModel", b =>
+            modelBuilder.Entity("Modules.Auth.UserModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
