@@ -1,7 +1,7 @@
 ﻿// <auto-generated />
 using System;
 using System.Collections.Generic;
-using Main.Data.Context;
+using Data.Context;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
@@ -9,7 +9,7 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 
 #nullable disable
 
-namespace Main.Data.Migrations
+namespace Data.Migrations
 {
     [DbContext(typeof(BaseContext))]
     partial class BaseContextModelSnapshot : ModelSnapshot
@@ -24,7 +24,7 @@ namespace Main.Data.Migrations
 
             NpgsqlModelBuilderExtensions.UseIdentityByDefaultColumns(modelBuilder);
 
-            modelBuilder.Entity("Main.Data.Models.GroupModel", b =>
+            modelBuilder.Entity("Data.Models.GroupModel", b =>
                 {
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd()
@@ -36,7 +36,7 @@ namespace Main.Data.Migrations
                     b.Property<Guid?>("AuthorId")
                         .HasColumnType("uuid");
 
-                    b.Property<string>("Descriere")
+                    b.Property<string>("Description")
                         .HasColumnType("text");
 
                     b.Property<DateTimeOffset?>("ExecutionDate")
@@ -56,7 +56,7 @@ namespace Main.Data.Migrations
                     b.ToTable("Groups", "base");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.LinkModel", b =>
+            modelBuilder.Entity("Data.Models.LinkModel", b =>
                 {
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd()
@@ -102,7 +102,7 @@ namespace Main.Data.Migrations
                     b.ToTable("Links", "base");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.NotificationModel", b =>
+            modelBuilder.Entity("Data.Models.MessageModel", b =>
                 {
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd()
@@ -117,6 +117,11 @@ namespace Main.Data.Migrations
                     b.Property<Guid?>("AuthorId")
                         .HasColumnType("uuid");
 
+                    b.Property<string>("Discriminator")
+                        .IsRequired()
+                        .HasMaxLength(13)
+                        .HasColumnType("character varying(13)");
+
                     b.Property<DateTimeOffset?>("ExecutionDate")
                         .HasColumnType("timestamptz");
 
@@ -124,27 +129,31 @@ namespace Main.Data.Migrations
                         .IsRequired()
                         .HasColumnType("text");
 
-                    b.Property<DateTimeOffset?>("ReadDate")
+                    b.Property<DateTimeOffset?>("SeenDate")
                         .HasColumnType("timestamptz");
 
                     b.PrimitiveCollection<List<string>>("Tags")
                         .HasColumnType("text[]");
+
+                    b.Property<Guid>("To")
+                        .HasColumnType("uuid");
 
                     b.HasKey("Id");
 
                     b.HasIndex("AuthorDataId");
 
                     b.ToTable("Notifications", "base");
+
+                    b.HasDiscriminator().HasValue("MessageModel");
+
+                    b.UseTphMappingStrategy();
                 });
 
-            modelBuilder.Entity("Main.Data.Models.PostModel", b =>
+            modelBuilder.Entity("Data.Models.ParameterModel", b =>
                 {
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd()
                         .HasColumnType("uuid");
-
-                    b.PrimitiveCollection<List<Guid>>("Attachments")
-                        .HasColumnType("uuid[]");
 
                     b.Property<Guid?>("AuthorDataId")
                         .HasColumnType("uuid");
@@ -152,30 +161,27 @@ namespace Main.Data.Migrations
                     b.Property<Guid?>("AuthorId")
                         .HasColumnType("uuid");
 
+                    b.Property<string>("Description")
+                        .HasColumnType("text");
+
                     b.Property<DateTimeOffset?>("ExecutionDate")
                         .HasColumnType("timestamptz");
 
-                    b.Property<string>("Message")
+                    b.Property<string>("Name")
                         .IsRequired()
                         .HasColumnType("text");
 
-                    b.Property<DateTimeOffset?>("SentDate")
-                        .HasColumnType("timestamp with time zone");
-
                     b.PrimitiveCollection<List<string>>("Tags")
                         .HasColumnType("text[]");
-
-                    b.Property<Guid?>("To")
-                        .HasColumnType("uuid");
 
                     b.HasKey("Id");
 
                     b.HasIndex("AuthorDataId");
 
-                    b.ToTable("Posts", "base");
+                    b.ToTable("Parameters", "base");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.QuestionModel", b =>
+            modelBuilder.Entity("Data.Models.QuestionModel", b =>
                 {
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd()
@@ -213,7 +219,7 @@ namespace Main.Data.Migrations
                     b.ToTable("Questions", "base");
                 });
 
-            modelBuilder.Entity("Main.Modules.Audit.AuditModel", b =>
+            modelBuilder.Entity("Modules.Audit.AuditModel", b =>
                 {
                     b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
@@ -255,7 +261,7 @@ namespace Main.Data.Migrations
                     b.ToTable("Audit", "base");
                 });
 
-            modelBuilder.Entity("Main.Modules.Auth.UserModel", b =>
+            modelBuilder.Entity("Modules.Auth.UserModel", b =>
                 {
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd()
@@ -316,63 +322,70 @@ namespace Main.Data.Migrations
                     b.ToTable("Users", "base");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.GroupModel", b =>
+            modelBuilder.Entity("Data.Models.PostModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasBaseType("Data.Models.MessageModel");
+
+                    b.HasDiscriminator().HasValue("PostModel");
+                });
+
+            modelBuilder.Entity("Data.Models.GroupModel", b =>
+                {
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
                     b.Navigation("AuthorData");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.LinkModel", b =>
+            modelBuilder.Entity("Data.Models.LinkModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
                     b.Navigation("AuthorData");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.NotificationModel", b =>
+            modelBuilder.Entity("Data.Models.MessageModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
                     b.Navigation("AuthorData");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.PostModel", b =>
+            modelBuilder.Entity("Data.Models.ParameterModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
                     b.Navigation("AuthorData");
                 });
 
-            modelBuilder.Entity("Main.Data.Models.QuestionModel", b =>
+            modelBuilder.Entity("Data.Models.QuestionModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
                     b.Navigation("AuthorData");
                 });
 
-            modelBuilder.Entity("Main.Modules.Audit.AuditModel", b =>
+            modelBuilder.Entity("Modules.Audit.AuditModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
                     b.Navigation("AuthorData");
                 });
 
-            modelBuilder.Entity("Main.Modules.Auth.UserModel", b =>
+            modelBuilder.Entity("Modules.Auth.UserModel", b =>
                 {
-                    b.HasOne("Main.Modules.Auth.UserModel", "AuthorData")
+                    b.HasOne("Modules.Auth.UserModel", "AuthorData")
                         .WithMany()
                         .HasForeignKey("AuthorDataId");
 
