@@ -1,14 +1,7 @@
-using System.ComponentModel.DataAnnotations.Schema;
+namespace Data.Models;
 
-namespace Main.Data.Models
+public class NotificationModel : MessageModel
 {
-    public class NotificationModel : BaseModel<Guid>
-    {
-        public required string Message { get; set; }
-
-        [Column(TypeName = "timestamptz")]
-        public DateTimeOffset? ReadDate { get; set; } = null;
-
-        public required List<Guid>? Attachments { get; set; } = null;
-    }
+    public required string Title { get; set; }
+    public string? Link { get; set; }
 }
