@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
 
-namespace Main.Modules.Auth;
+namespace Modules.Auth;
 
 public class RegistrationDTO
 {
